Backend/requirements.txt
flask
flask-cors
gunicorn
numpy
pandas
scikit-learn
tensorflow-cpu
requests
